+++ InternetShop/src/main/java/service/ProductServiceImpl.java
@@ -1,5 +1,4 @@
 package service;
 
 public class ProductServiceImpl implements ProductService {
-
 }
