--- InternetShop/src/main/java/service/RegisterService.java
@@ -1,5 +1,0 @@
-package service;
-
-public interface RegisterService {
-    boolean registration (String name, String login, String password, String email);
-}
